requirements.txt
Flask==3.0.3
PyYAML==6.0.2
gunicorn==22.0.0